move VM_DSET_READING_ORDER_RTL to viewmap dataset record per C-API
The name of `READING_ORDER_RTL` is VM_DSET_ READING_ORDER_RTL, suggesting that it's a dataset record flag value. Also, the C-API lists it as such. It's misleadingly grouped with the VM_DROBJ flag values in the CAPI .h file. So, i moved that flag to viewmapdatasetrecord. (In any case it seems it was the wrong value, should be 0x0020, not 0x0010).